Create VRS class for creating VRS objects
