+++ variation/validators/coding_dna_deletion.py
@@ -65,7 +65,7 @@ class CodingDNADeletion(DuplicationDeletionBase):
                 if cds_start_end is not None:
                     cds_start = cds_start_end[0]
 
-                    allele = self.to_vrs_allele(
+                    allele = self.vrs.to_vrs_allele(
                         t, s.start_pos_del, s.end_pos_del,
                         s.reference_sequence, s.alt_type,
                         errors, cds_start=cds_start
