+++ variation/validators/single_nucleotide_variation_base.py
@@ -139,7 +139,7 @@ class SingleNucleotideVariationBase(Validator):
                     cds_start = None
 
                 if not errors:
-                    allele = self.to_vrs_allele(
+                    allele = self.vrs.to_vrs_allele(
                         t, s.position, s.position, s.reference_sequence,
                         s.alt_type, errors, cds_start=cds_start,
                         alt=s.new_nucleotide
