+++ variation/validators/validator.py
@@ -1,6 +1,6 @@
 """Module for Validation."""
 import copy
-from typing import List, Optional, Dict, Tuple, Union
+from typing import List, Optional, Dict, Tuple
 from abc import ABC, abstractmethod
 from variation.schemas.classification_response_schema import Classification, \
     ClassificationType
@@ -15,14 +15,12 @@ from ga4gh.vrs.dataproxy import SeqRepoDataProxy
 from ga4gh.vrs.extras.translator import Translator
 import hgvs.parser
 import logging
-from ga4gh.vrs import models, normalize
-from ga4gh.core import ga4gh_identify
 from variation.validators.genomic_base import GenomicBase
 from variation.data_sources import UTA
-from bioutils.accessions import coerce_namespace
 from gene.query import QueryHandler as GeneQueryHandler
 from variation.schemas.normalize_response_schema\
     import HGVSDupDelMode as HGVSDupDelModeEnum
+from variation.vrs import VRS
 
 logger = logging.getLogger('variation')
 logger.setLevel(logging.DEBUG)
@@ -36,7 +34,8 @@ class Validator(ABC):
                  gene_symbol: GeneSymbol,
                  mane_transcript: MANETranscript,
                  uta: UTA, dp: SeqRepoDataProxy, tlr: Translator,
-                 gene_normalizer: GeneQueryHandler) -> None:
+                 gene_normalizer: GeneQueryHandler,
+                 vrs: VRS) -> None:
         """Initialize the DelIns validator.
 
         :param SeqRepoAccess seqrepo_access: Access to SeqRepo data
@@ -48,6 +47,7 @@ class Validator(ABC):
         :param UTA uta: Access to UTA queries
         :param Translator tlr: Translator class
         :param GeneQueryHandler gene_normalizer: Access to gene-normalizer
+        :param VRS vrs: Class for creating VRS objects
         """
         self.transcript_mappings = transcript_mappings
         self.seqrepo_access = seqrepo_access
@@ -59,6 +59,7 @@ class Validator(ABC):
         self.genomic_base = GenomicBase(self.dp, self.uta)
         self.mane_transcript = mane_transcript
         self.gene_normalizer = gene_normalizer
+        self.vrs = vrs
 
     @abstractmethod
     def is_token_instance(self, t: Token) -> bool:
@@ -320,9 +321,9 @@ class Validator(ABC):
         return [t for t in classification.all_tokens
                 if self.is_token_instance(t)]
 
+    @staticmethod
     def get_gene_symbol_tokens(
-            self, classification: Classification
-    ) -> List[Optional[GeneMatchToken]]:
+            classification: Classification) -> List[Optional[GeneMatchToken]]:
         """Return tokens with GeneSymbol token type from a classification.
 
         :param Classification classification: Classification of input string
@@ -411,7 +412,8 @@ class Validator(ABC):
         ]
         return self._get_gene_tokens(classification, mappings)
 
-    def get_accession(self, t: str, classification: Classification) -> str:
+    @staticmethod
+    def get_accession(t: str, classification: Classification) -> str:
         """Return accession for a classification
 
         :param str t: Accession
@@ -475,207 +477,6 @@ class Validator(ABC):
             )
             return False
 
-    def _get_start_indef_range(self, start: int) -> models.IndefiniteRange:
-        """Return indefinite range given start coordinate
-
-        :param int start: Start position (assumes 1-based)
-        :return: Indefinite range model
-        """
-        return models.IndefiniteRange(value=start - 1, comparator="<=")
-
-    def _get_end_indef_range(self, end: int) -> models.IndefiniteRange:
-        """Return indefinite range given end coordinate
-
-        :param int end: End position (assumes 1-based)
-        :return: Indefinite range model
-        """
-        return models.IndefiniteRange(value=end, comparator=">=")
-
-    def _get_ival_certain_range(self, start1: int, start2: int, end1: int,
-                                end2: int) -> models.SequenceInterval:
-        """Return sequence interval
-
-        :param int start1: Start left pos (assumes 1-based)
-        :param int start2: Start right pos (assumes 1-based)
-        :param int end1: End left pos (assumes 1-based)
-        :param int end2: End right pos (assumes 1-based)
-        :return: Sequence Interval model
-        """
-        return models.SequenceInterval(
-            start=models.DefiniteRange(min=start1 - 1, max=start2 - 1),
-            end=models.DefiniteRange(min=end1 + 1, max=end2 + 1)
-        )
-
-    def _get_sequence_loc(
-            self, ac: str, interval: models.SequenceInterval
-    ) -> models.Location:
-        """Return VRS location
-
-        :param str ac: Accession
-        :param models.SequenceInterval interval: VRS sequence interval
-        :return: VRS Location model
-        """
-        return models.Location(sequence_id=coerce_namespace(ac),
-                               interval=interval)
-
-    def _get_ival_start_end(
-            self, coordinate: str, start: int, end: int, cds_start: int,
-            errors: List) -> Optional[Tuple[int, int]]:
-        """Get ival_start and ival_end coordinates.
-
-        :param str coordinate: Coordinate used. Must be either `p`, `c`, or `g`
-        :param int start: Start position change
-        :param int end: End position change
-        :param int cds_start: Coding start site
-        :param List errors: List of errors
-        :return: Tuple[ival_start, ival_end]
-        """
-        try:
-            start = int(start)
-            if end is None:
-                end = start
-            end = int(end)
-        except (ValueError, TypeError):
-            errors.append("Start/End must be valid ints")
-            return None
-
-        if coordinate == 'c':
-            if cds_start:
-                start += cds_start
-                end += cds_start
-        return start, end
-
-    def to_vrs_allele_ranges(
-            self, ac: str, coordinate: str, alt_type: str, errors: List,
-            ival: models.SequenceInterval) -> Optional[Dict]:
-        """Translate variation ranges to VRS Allele Object.
-
-        :param str ac: Accession
-        :param str coordinate: Coordinate used. Must be either `p`, `c`, or `g`
-        :param str alt_type: Type of alteration
-        :param list errors: List of errors
-        :param models.SequenceInterval ival: Sequence Interval
-        :return: VRS Allele object
-        """
-        if coordinate == 'c':
-            # TODO: Once we add support for ranges on c. coord
-            return None
-        if alt_type in ['uncertain_deletion', 'uncertain_duplication',
-                        'duplication_range', 'deletion_range']:
-            sstate = models.LiteralSequenceExpression(
-                sequence=""
-            )
-        else:
-            errors.append("No state")
-            return None
-
-        return self._vrs_allele(ac, ival, sstate, alt_type, errors)
-
-    def _vrs_allele(self, ac: str, interval: models.SequenceInterval,
-                    sstate: Union[models.LiteralSequenceExpression,
-                                  models.DerivedSequenceExpression,
-                                  models.RepeatedSequenceExpression],
-                    alt_type: str, errors: List) -> Optional[Dict]:
-        """Create a VRS Allele object.
-
-        :param str ac: Accession
-        :param SequenceInterval interval: Sequence Interval
-        :param sstate: State
-        :type sstate: models.LiteralSequenceExpression or
-            models.DerivedSequenceExpression or
-            models.RepeatedSequenceExpression
-        :param str alt_type: Type of alteration
-        :param List errors: List of errors
-        :return: VRS Allele object represented as a Dict
-        """
-        location = self._get_sequence_loc(ac, interval)
-        allele = models.Allele(location=location, state=sstate)
-
-        # Ambiguous regions do not get normalized
-        if alt_type not in ["uncertain_deletion", "uncertain_duplication",
-                            "duplication_range", "deletion_range"]:
-            try:
-                allele = normalize(allele, self.dp)
-                if alt_type == 'deletion':
-                    allele.state.sequence = ''
-            except (KeyError, AttributeError) as e:
-                errors.append(f"vrs-python unable to normalize allele: {e}")
-                return None
-
-        if not allele:
-            errors.append("Unable to get allele")
-            return None
-
-        seq_id = self.dp.translate_sequence_identifier(
-            allele.location.sequence_id._value, "ga4gh")[0]
-        allele.location.sequence_id = seq_id
-        allele.location._id = ga4gh_identify(allele.location)
-        allele._id = ga4gh_identify(allele)
-        return allele.as_dict()
-
-    def to_vrs_allele(
-            self, ac: str, start: int, end: int, coordinate: str,
-            alt_type: str, errors: List, cds_start: int = None,
-            alt: str = None) -> Optional[Dict]:
-        """Translate accession and position to VRS Allele Object.
-
-        :param str ac: Accession
-        :param int start: Start position change
-        :param int end: End position change
-        :param str coordinate: Coordinate used. Must be either `p`, `c`, or `g`
-        :param str alt_type: Type of alteration
-        :param List errors: List of errors
-        :param int cds_start: Coding start site
-        :param str alt: Alteration
-        :return: VRS Allele Object
-        """
-        ival_coords = self._get_ival_start_end(coordinate, start, end,
-                                               cds_start, errors)
-        if not ival_coords:
-            return None
-        if ival_coords[0] > ival_coords[1]:
-            ival_end, ival_start = ival_coords
-        else:
-            ival_start, ival_end = ival_coords
-
-        # Right now, this follows HGVS conventions
-        # This will change once we support other representations
-        if alt_type == 'insertion':
-            state = alt
-            ival_end = ival_start
-        elif alt_type in ['substitution', 'deletion', 'delins',
-                          'silent_mutation', 'nonsense']:
-            if alt_type == 'silent_mutation':
-                state = self.seqrepo_access.get_sequence(
-                    ac, ival_start
-                )
-                if state is None:
-                    errors.append(f"Unable to get sequence on {ac} from "
-                                  f"{ival_start}")
-                    return None
-            else:
-                state = alt or ''
-            ival_start -= 1
-        elif alt_type == 'duplication':
-            ref = self.seqrepo_access.get_sequence(ac, ival_start,
-                                                   ival_end)
-            if ref is not None:
-                state = ref + ref
-            else:
-                errors.append(f"Unable to get sequence on {ac} from "
-                              f"{ival_start} to {ival_end}")
-                return None
-            ival_start -= 1
-        else:
-            errors.append(f"alt_type not supported: {alt_type}")
-            return None
-
-        interval = models.SequenceInterval(
-            start=models.Number(value=ival_start),
-            end=models.Number(value=ival_end))
-        sstate = models.LiteralSequenceExpression(sequence=state)
-        return self._vrs_allele(ac, interval, sstate, alt_type, errors)
-
     def _validate_gene_pos(self, gene: str, alt_ac: str, pos1: int, pos2: int,
                            errors: List, pos3: int = None, pos4: int = None,
                            residue_mode: str = "residue") -> None:
@@ -732,7 +533,8 @@ class Validator(ABC):
                         errors.append(f"Position {pos} out of index on "
                                       f"{alt_ac} on gene, {gene}")
 
-    def _get_coord_alt(self, coordinate: str, mane: Dict,
+    @staticmethod
+    def _get_coord_alt(coordinate: str, mane: Dict,
                        s_copy: Token) -> Optional[Tuple[str, str]]:
         """Get coordinate and alteration
 
@@ -793,7 +595,7 @@ class Validator(ABC):
             alt = coord_alt[1] if coord_alt[1] else alt
 
         if mane_variation is None:
-            new_allele = self.to_vrs_allele(
+            new_allele = self.vrs.to_vrs_allele(
                 mane['refseq'], mane['pos'][0], mane['pos'][1],
                 coordinate, alt_type, [],
                 cds_start=mane.get('coding_start_site', None), alt=alt
@@ -808,7 +610,8 @@ class Validator(ABC):
         self._add_dict_to_mane_data(mane['refseq'], s_copy, variation,
                                     mane_data, mane['status'])
 
-    def _add_dict_to_mane_data(self, ac: str, s: Token, variation: Dict,
+    @staticmethod
+    def _add_dict_to_mane_data(ac: str, s: Token, variation: Dict,
                                mane_data: Dict, status: str) -> None:
         """Add variation data to mane data for normalize endpoint.
 
@@ -886,7 +689,8 @@ class Validator(ABC):
             return None
         return seq
 
-    def _grch38_dict(self, ac: str, pos: Tuple[int, int]) -> Dict:
+    @staticmethod
+    def _grch38_dict(ac: str, pos: Tuple[int, int]) -> Dict:
         """Create dict for normalized concepts
 
         :param str ac: Acession
